grams per meal

+++ SztucznaIntCw/SztucznaIntCw/Classes/Interfaces/IMeal.cs
@@ -4,10 +4,10 @@ namespace SztucznaIntCw.Classes.Interfaces
 {
     public interface IMeal
     {
-       
-        int TotalGramsOfProteins { get; set; }
-        int TotalGramsOfFats { get; set; }
-        int TotalGramsOfCarbs { get; set; }
+
+        decimal TotalGramsOfProteins { get; set; }
+        decimal TotalGramsOfFats { get; set; }
+        decimal TotalGramsOfCarbs { get; set; }
 
         List<IProduct> ProductList { get; }
     }
