+++ SztucznaIntCw/SztucznaIntCw/Classes/NonAbstract/Meal.cs
@@ -5,9 +5,9 @@ namespace SztucznaIntCw.Classes.NonAbstract
 {
     public class Meal : IMeal
     {
-        public int TotalGramsOfProteins { get; set; }
-        public int TotalGramsOfFats { get; set; }
-        public int TotalGramsOfCarbs { get; set; }
+        public decimal TotalGramsOfProteins { get; set; }
+        public decimal TotalGramsOfFats { get; set; }
+        public decimal TotalGramsOfCarbs { get; set; }
 
         public List<IProduct> ProductList { get;  set; }
